add commit

+++ libs/VS2022/render/application/Engine.cpp
@@ -51,7 +51,7 @@ void ll::Engine::SetPreferredFramesPerSecond(int fps)
 {
 }
 
-ll::uint ll::Engine::GetTotalFrames() const
+uint32_t ll::Engine::GetTotalFrames() const
 {
     return _totalFrames;
 }
